Project 2.0/requirements.txt
streamlit
sqlalchemy
mysql-connector-python
pydantic
python-dotenv
pymysql
requests
pytest
streamlit-option-menu
pandas
altair
